- Adding initial health system for player

+++ Assets/Scripts/Constants.cs
@@ -12,7 +12,7 @@ public static class Constants
     /// The green sword is considered the base attack power for the game... it's 2 half hearts, so 1 whole heart is its damage
     /// Each sword then has a modifier based on this value... White Sword's modifier is 2x, and Magical Sword is 4x
     /// </summary>
-    public const int BASE_SWORD_POWER = 2;
+    public const float BASE_SWORD_POWER = 2f;
 
     public static readonly Vector3 SWORD_RIGHT = new Vector3(0.11f, -0.01f, 0f);
     public static readonly Vector3 SWORD_LEFT = new Vector3(-0.11f, -0.01f, 0f);
